requirements.txt
# AI Providers
anthropic>=0.40.0
groq>=0.11.0

# Telegram Bot
python-telegram-bot>=21.0

# Scheduling
apscheduler>=3.10.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0

# HTTP Client
httpx>=0.27.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Indian Stock Market (Zerodha)
kiteconnect>=5.0.0
pyotp>=2.9.0

# US Stock Market (Alpaca)
alpaca-py>=0.21.0

# Crypto Trading
ccxt>=4.2.0
freqtrade>=2024.1

# Historical Data (FREE - no Zerodha paid plan needed)
yfinance>=0.2.36

# Technical Analysis
pandas-ta>=0.3.14b
mplfinance>=0.12.10

# Chart Generation
matplotlib>=3.8.0

# News APIs
finnhub-python>=2.4.19
praw>=7.7.0

# Backtesting
backtrader>=1.9.78

# Market Holiday Calendars (dynamic - NSE, NYSE, etc.)
exchange_calendars>=4.5.0

# WebSocket (real-time Binance streaming)
websockets>=12.0

# Encryption (Fernet for secrets at rest)
cryptography>=42.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
